Refactor DayOfWeek handling with enum implementation
This commit introduces a new enum type, `DayOfWeekEnum`, to replace the integer representation of `DayOfWeekId` in various entities. The `DayOfWeekId` property has been changed from an `int` to a `byte` in the `BusinessHoursEntity` and `DayOfWeekEntity` classes.

Migration files have been updated to reflect these changes, ensuring the database schema aligns with the new logic. Data seeding for business hours and day of the week has also been updated to utilize the new enum values.

These changes improve code clarity, maintainability, and type safety.

+++ src/Application/Domain/Entities/BusinessHoursEntity.cs
@@ -18,7 +18,7 @@ public class BusinessHoursEntity
 
     [Column("day_of_week_id")]
     [Required]
-    public int DayOfWeekId { get; set; }
+    public byte DayOfWeekId { get; set; }
 
     [Column("open_time")]
     public LocalTime? OpenTime { get; set; }
